+++ controller/src/main/java/com/linbit/linstor/event/handler/protobuf/controller/DonePercentageEventHandler.java
@@ -1,6 +1,9 @@
 package com.linbit.linstor.event.handler.protobuf.controller;
 
 import com.linbit.linstor.InternalApiConsts;
+import com.linbit.linstor.annotation.SystemContext;
+import com.linbit.linstor.core.identifier.NodeName;
+import com.linbit.linstor.core.repository.NodeRepository;
 import com.linbit.linstor.event.EventIdentifier;
 import com.linbit.linstor.event.common.DonePercentageEvent;
 import com.linbit.linstor.event.handler.EventHandler;
@@ -8,7 +11,10 @@ import com.linbit.linstor.event.handler.SatelliteStateHelper;
 import com.linbit.linstor.event.handler.protobuf.ProtobufEventHandler;
 import com.linbit.linstor.proto.eventdata.EventDonePercentageOuterClass;
 import com.linbit.linstor.satellitestate.SatelliteVolumeState;
+import com.linbit.linstor.security.AccessContext;
 import com.linbit.utils.Pair;
+
+import static com.linbit.linstor.event.handler.protobuf.controller.ReplicationStateEventHandler.getMappedName;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -25,15 +31,21 @@ public class DonePercentageEventHandler implements EventHandler
 {
     private final SatelliteStateHelper satelliteStateHelper;
     private final DonePercentageEvent donePercentageEvent;
+    private final AccessContext sysCtx;
+    private final NodeRepository nodeRepo;
 
     @Inject
     public DonePercentageEventHandler(
         SatelliteStateHelper satelliteStateHelperRef,
-        DonePercentageEvent donePercentageEventRef
+        DonePercentageEvent donePercentageEventRef,
+        @SystemContext AccessContext sysCtxRef,
+        NodeRepository nodeRepositoryRef
     )
     {
         satelliteStateHelper = satelliteStateHelperRef;
         donePercentageEvent = donePercentageEventRef;
+        sysCtx = sysCtxRef;
+        nodeRepo = nodeRepositoryRef;
     }
 
     @Override
@@ -47,9 +59,10 @@ public class DonePercentageEventHandler implements EventHandler
             EventDonePercentageOuterClass.EventDonePercentage eventDonePercentage =
                 EventDonePercentageOuterClass.EventDonePercentage.parseDelimitedFrom(eventDataIn);
 
+            NodeName mappedName = getMappedName(nodeRepo, sysCtx, eventDonePercentage.getPeerName());
             donePercentage = eventDonePercentage.hasDonePercentage() ?
-                new Pair<>(eventDonePercentage.getPeerName(), Optional.of(eventDonePercentage.getDonePercentage())) :
-                new Pair<>(eventDonePercentage.getPeerName(), Optional.empty());
+                new Pair<>(mappedName.displayValue, Optional.of(eventDonePercentage.getDonePercentage())) :
+                new Pair<>(mappedName.displayValue, Optional.empty());
             satelliteStateHelper.onSatelliteState(
                 eventIdentifier.getNodeName(),
                 satelliteState -> satelliteState.setOnVolume(